Admin: Add support for host-based checking.
Arguably more secure, may prevent spoofing in some cases where a user changes there nick, issues something pre-auth, then their nick changes back. With host-based, the user must be using the same host as you, which may not be likely if you're using a bouncer and they're not, as an example, or completely different locations.

A follow-up commit will introduce the ability to disable specific admin commands entirely for the most extensive level of security.